BackEnd/requirements.txt
bcrypt==4.0.1
db-dtypes==1.1.1
Fiona==1.9.4
Flask==2.3.2
Flask-Cors==3.0.10
geopandas==0.13.0
matplotlib==3.7.1
pandas==2.0.1
PyJWT==2.7.0
pyproj==3.5.0
python-dotenv==1.0.0
rasterio==1.3.6
requests==2.30.0
scipy==1.10.1
shapely==2.0.1
statsmodels==0.14.0
